+++ app/policies/level_b_policy.rb
@@ -6,4 +6,12 @@ class LevelBPolicy < ApplicationPolicy
   def budget_upload?
     return true if beis_user?
   end
+
+  def create_activity_comment?
+    update_activity_comment?
+  end
+
+  def update_activity_comment?
+    return true if beis_user?
+  end
 end